refactor (Sidebar): Move component to shared and improve folder structure
