+++ src/ui/server-administration/shared/Sidebar/Sidebar.scss
@@ -1,5 +1,5 @@
 @import "~realm-studio-styles/variables";
-@import "../variables";
+@import "../../variables";
 
 .Sidebar {
   background: $white;
