+++ src/ui/server-administration/shared/Sidebar/SidebarTable.scss
