--- src/ui/server-administration/sidebar/index.tsx
@@ -1,1 +1,0 @@
-export { default as Sidebar, IProps as ISidebar } from './Sidebar';
